Requirement 3 changes. Visuals The application should display
1. a list of at least 5 cats, listed by name
2. an area to display the selected cat
3. In the cat display area, the following should be displayed

a. the cat's name
b. a picture of the cat
c. text showing the number of clicks
d. The specifics of the layout do not matter, so style it however you'd like.
Interaction
When a cat name is clicked in the list, the cat display area should update to show the data for the selected cat.
The number of clicks in the cat area should be unique to each cat, and should increment when the cat's picture is clicked.